+++ src/main/java/yangchen/exam/repo/examInfoRepo.java
@@ -9,21 +9,21 @@ import java.util.List;
 
 public interface examInfoRepo extends JpaRepository<ExamInfo, Integer> {
 
-    List<ExamInfo> findByStudentNumber(Long studentId);
+    List<ExamInfo> findByStudentNumber(Integer studentId);
 
     ExamInfo findByExaminationId(Integer examinationId);
 
     List<ExamInfo> findByExamGroupId(Integer examGroupId);
 
 
-    List<ExamInfo> findByStudentNumberAndExamStartAfter(Long studentId, Timestamp timestamp);
+    List<ExamInfo> findByStudentNumberAndExamStartAfter(Integer studentId, Timestamp timestamp);
 
-    List<ExamInfo> findByStudentNumberAndExamStartBeforeAndExamEndAfter(Long studentId, Timestamp timestamp, Timestamp endTime);
+    List<ExamInfo> findByStudentNumberAndExamStartBeforeAndExamEndAfter(Integer studentId, Timestamp timestamp, Timestamp endTime);
 
-    List<ExamInfo> findByStudentNumberAndExamEndBefore(Long studentId, Timestamp timestamp);
+    List<ExamInfo> findByStudentNumberAndExamEndBefore(Integer studentId, Timestamp timestamp);
 
     @Query(value = "select  * from exam_info join examation on exam_info.examination_id=examation.id\n" +
             "where examation.used=1 and  student_number=?", nativeQuery = true)
-    List<ExamInfo> getFinishedExam(Long studentId);
+    List<ExamInfo> getFinishedExam(Integer studentId);
 
 }
